lint

+++ apps/main-landing/src/app/vault/commands/get-staked-balance.ts
@@ -20,7 +20,7 @@ const getStakedBalance = async (payload: Payload) => {
     args: [payload.address],
   });
 
-  return formatEther(balance as bigint) ?? '0';
+  return formatEther(balance) ?? '0';
 };
 
 export const useGetStakedBalance = (payload: Payload) => {
